+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added an ability to prepare rows before appending rows to sheet. Just add `prepareRows` method for your export class if needed.
+- Added ab ability to catch exceptions from `QueueExport` job. Just add `failed` method for your export class if needed.
 
 ## [3.1.24] - 2020-10-28